fix disk/raid handling with partprobe
partprobe might actually trigger an mdadm scan with automatic assemble
of raid partitions/devices.
This will cause subsequent operations on the disks partition table
and/or partition to fail and leave the kernel stuck with the old
partition table.

This commit fixes the mdadm/dmraid stop function stop_lvm_raid() and
replaces varying occurances of manual stop procedurs with a call to this
method.
Additionall it makes sure that after each partprobe this stop_lvm_raid()
method is called.

Signed-off-by: Thore Bödecker <me@foxxx0.de>