enforce lvm commands to be non-interactive
From time to time lvcreate calls would fail if the PE boundary of the
newly created LV resulted in a filesystem header (e.g. xfs/ext4) being
at the beginning of it.
In such cases lvcreate would prompt the user interactively if these
signatures should be overwritten or not.
Obviously this is not desired in such an automated scripting sitation,
therefore I'm adding the required parameters for the
pvcreate/vgcreate/lvcreate calls so that they automatically overwrite
exisiting headers/signatures and assummes "yes" by default.

Signed-off-by: Thore Bödecker <me@foxxx0.de>