only use check_dos_partitions without GPT
Unfortunately the function check_dos_partitions() not only checks, but
also changes certain variables, its use should be limited to where
absolutely necessary.
This method calculates an absolute nuumerical value for the special
"all" partition size, which breaks later special handling of the "all"
value, as it is no longer present.

Within the GPT/sgdisk code there is a case for dealing with the "all"
parititon where it uses `--largest-new` for "all" instead of an absolute
value, which removes the need to manually calculate the partition size.

Signed-off-by: Thore Bödecker <me@foxxx0.de>