fix FORCE_GPT parsing and exporting
Previosly the grep command in question could also result in an empty
string if FORCE_GPT was not specified in the provided installimage
config.

This commit checks if the FORCE_GPT param was actually present in the
config before trying to overwrite it.
Additionally, it is initialized to "1" if it was an emptystring
beforehand.
You can now use your local config.sh to export a value for FORCE_GPT, as
in this case the initialization to "1" would not take place.

Furthermore the variable FORCE_GPT is now properly exported in the
correct place, so it present in the env while install.sh and
autosetup.sh are running.

Signed-off-by: Thore Bödecker <me@foxxx0.de>