add ability to enable/disable dwc
The so-called DWC (drive write cache) option is a feature supported by
most modern hard disks and solid state drives.
It allows the respective drive to cache write requests in the internal
drive-cache before flushing it to the actual persistent memory.
This implies, that a sync() is not necessarily already written to disk
and thus might cause filesystem corruption.

Depending on OS and kernel versions as well as type of hardware, this
setting varies between not supported, off and on.

On newer kernels and SSDs this is mostly enabled by default, as the DWC
is very helpful to reduce write amplification and throughput on NAND
flash drives by allowing the drive to queue requests and always write
full blocks.

The config option "DWC_ENABLE" introduced with this commit defaults to 1
in a similar manner as "FORCE_GPT".
A setting of 0 means that DWC will be forced off for all drives.
A setting of 1 means that DWC will be enabled for SSDs but not HDDs.
A setting of 2 means that DWC will be enabled for all drives.

Signed-off-by: Thore Bödecker <me@foxxx0.de>